+++ src/utils/index.js
@@ -132,9 +132,7 @@ const removeDiacritics = str => {
     { base: 'z', letters: /[\u007A\u24E9\uFF5A\u017A\u1E91\u017C\u017E\u1E93\u1E95\u01B6\u0225\u0240\u2C6C\uA763]/g }
   ]
 
-  return defaultDiacriticsRemovalMap.reduce((string, { letters, base }) => {
-    return string.replace(letters, base)
-  }, str)
+  return defaultDiacriticsRemovalMap.reduce((string, { letters, base }) => string.replace(letters, base), str)
 }
 
 const removeUndefined = obj => {