+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 ## Changelog
 
+### 2.1.0
+
+* **New:** Added vue3 and bootstrap5 support.
+* **New:** Added lint for js, scss, spell and editor.
+* **New:** Added `classInput` option to show class of input radio/checkbox.
+* **New:** Added `filterSelectAll` option to show/hide the select all checkbox in the filter.
+* **New:** Added `resetFilter` method to reset the filter.
+* **New:** Added id-ID locale.
+* **Update:** Updated `placeholder` class to `ms-placeholder`.
+* **Update:** Fixed alignment issue of single mode in the bootstrap theme.
+
 ### 2.0.9
 
 - **New:** Added `vue3` support.
@@ -200,7 +211,7 @@
 * [bug] Fix #45, #64: update width option to support a percentage setting.
 * [bug] Trigger the checkbox on the entire line.
 * [bug] Added `noMatchesFound` option.
-* [bug] Update `seperator` to `separator`.
+* [bug] Update `separator` typo.
 * [enh] Allow object of options along with initial method.
 * [enh] Add a `filterAcceptOnEnter` option.
 * [enh] Put class on ms-parent div instead of ul.
@@ -259,7 +270,7 @@
 
 ### 1.0.5
 
-* Update the button text witdh.
+* Update the button text width.
 * Add keyboard support.
 
 ### 1.0.4